requirements.txt
beautifulsoup4==4.8.2
bs4==0.0.1
certifi==2019.11.28
chardet==3.0.4
Click==7.0
cssselect==1.1.0
feedfinder2==0.0.4
feedparser==5.2.1
Flask==1.1.1
Flask-SQLAlchemy==2.4.1
idna==2.8
itsdangerous==1.1.0
jieba3k==0.35.1
Jinja2==2.11.0
lxml==4.4.2
MarkupSafe==1.1.1
newspaper3k==0.2.8
nltk==3.4.5
Pillow==7.0.0
pkg-resources==0.0.0
python-dateutil==2.8.1
pytz==2019.3
PyYAML==5.3
requests==2.22.0
requests-file==1.4.3
six==1.14.0
soupsieve==1.9.5
SQLAlchemy==1.3.13
tinysegmenter==0.3
tldextract==2.2.2
tzlocal==2.0.0
urllib3==1.25.8
Werkzeug==0.16.1
